ra-barrier/requirements.txt
astroid==2.4.1
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Flask==1.1.2
gunicorn==20.0.4
imageio==2.8.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
networkx==2.4
numpy==1.18.5
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scipy==1.4.1
Shapely==1.7.0
six==1.15.0
tifffile==2020.6.3
toml==0.10.1
Werkzeug==1.0.1
wrapt==1.12.1
